back dev

+++ server/src/services/file.service.ts
@@ -47,12 +47,15 @@ export class FileService {
   }
 
   public async delete(id: string): Promise<File> {
-    try {
-       unlinkSync("./../data/images/"+id+'.jpg')
-     } catch(err) {
-       console.error(err)
-     }
-    const obj = await this.repository.delete(id)
+    const obj = await this.repository.findOne(id);
+    if (obj) {
+      try {
+        unlinkSync("./../data/files/"+id+"."+obj.name.split('.')[1])
+      } catch(err) {
+        console.error(err)
+      }
+     const del = await this.repository.delete(id)
+    }
     return 
   }
